The lab7 fully done, the work with SpringBoot is very interesting was written all logic for controllers and services, and inplemented GOF logic, also system was tested via JUNIT

+++ src/main/java/Lab7/controller/FlightController.java
@@ -2,6 +2,7 @@ package Lab7.controller;
 
 import Lab7.Flight;
 import Lab7.services.FlightService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -30,7 +31,13 @@ public class FlightController {
     }
 
     @GetMapping("/flights/new")
-    public String showCreateForm(Model model) {
+    public String showCreateForm(Model model, HttpSession session) {
+
+        String currentUser = (String) session.getAttribute("logged_in_user");
+
+        if (currentUser == null) {
+            return "redirect:/login";
+        }
 
         Flight flight = new Flight();
         model.addAttribute("flight", flight);
@@ -39,7 +46,13 @@ public class FlightController {
     }
 
     @PostMapping
-    public String saveFlight(@ModelAttribute Flight flight) {
+    public String saveFlight(@ModelAttribute Flight flight, HttpSession session) {
+
+        String currentUser = (String) session.getAttribute("logged_in_user");
+
+        if (currentUser == null) {
+            return "redirect:/login";
+        }
         // віддаємо заповнений об'єкт сервісу для збереження в базу
         flightService.save(flight);
 
@@ -49,7 +62,12 @@ public class FlightController {
 
 
     @GetMapping("/flights/delete/{id}")
-    public String deleteFlight(@PathVariable Long id) {
+    public String deleteFlight(@PathVariable Long id, HttpSession session) {
+
+        String currentUser = (String) session.getAttribute("logged_in_user");
+        if (currentUser == null) {
+            return "redirect:/login";
+        }
         flightService.deleteFlight(id);
 
         // повертаємо користувача назад до загальної таблиці
@@ -57,7 +75,13 @@ public class FlightController {
     }
 
     @GetMapping("/flights/edit/{id}")
-    public String showEditForm(@PathVariable Long id, Model model) {
+    public String showEditForm(@PathVariable Long id, Model model, HttpSession session) {
+
+        String currentUser = (String) session.getAttribute("logged_in_user");
+
+        if (currentUser == null) {
+            return "redirect:/login";
+        }
         Flight flight = flightService.getFlightById(id);
         model.addAttribute("flight", flight);
